# dist/lib/main/json/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Hyperledger Cactus Plugin - Connector Corda",
        "description": "Can perform basic tasks on a Corda ledger",
        "version": "v2.0.0-alpha.2",
        "license": {
            "name": "Apache-2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "components": {
        "schemas": {
...
            "ListFlowsV1Request": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "filter": {
                        "type": "string"
                    }
                }
            },
            "ListFlowsV1Response": {
                "type": "object",
                "required": ["flowNames"],
                "properties": {
                    "flowNames": {
                        "type": "array",
                        "description": "An array of strings storing the names of the flows as returned by the flowList Corda RPC.",
                        "example": [
                            "net.corda.finance.flows.CashExitFlow",
                            "net.corda.finance.flows.CashIssueAndPaymentFlow",
                            "net.corda.finance.flows.CashIssueFlow",
                            "net.corda.finance.flows.CashPaymentFlow",
                            "net.corda.finance.internal.CashConfigDataFlow",
                            "net.corda.samples.obligation.flows.IOUIssueFlow",
                            "net.corda.samples.obligation.flows.IOUSettleFlow",
                            "net.corda.samples.obligation.flows.IOUTransferFlow",
                            "net.corda.samples.obligation.flows.SelfIssueCashFlow"
                        ],
                        "items": {
                            "type": "string",
                            "minItems": 0,
                            "maxItems": 10000000
                        },
                        "default": []
                    }
                }
            },
...
            "StartFlowV5Request": {
                "description": "This method starts a new instance for the specified flow for the specified holding identity.",
                "type": "object",
                "required": ["clientRequestId", "flowClassName", "requestBody"],
                "properties": {
                    "holdingIDShortHash": {
                        "type": "string",
                        "nullable": false
                    },
                    "clientRequestId": {
                        "type": "string",
                        "nullable": false
                    },
                    "flowClassName": {
                        "type": "string",
                        "nullable": false
                    },
                    "requestBody": {
                        "type": "object",
                        "properties": {
                            "chatName": {
                                "type": "string"
                            },
                            "otherMember": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            },
                            "numberOfRecords": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "FlowStatusV5Response": {
                "type": "object",
                "required": ["flowStatus", "holdingIDShortHash", "timestamp"],
                "properties": {
                    "clientRequestId": {
                        "type": "string",
                        "nullable": true,
                        "example": "string"
                    },
                    "flowError": {
                        "$ref": "#/components/schemas/FlowV5Error"
                    },
                    "flowId": {
                        "type": "string",
                        "nullable": true,
                        "example": "string"
                    },
                    "flowResult": {
                        "type": "string",
                        "nullable": true,
                        "example": "string"
                    },
                    "flowStatus": {
                        "type": "string",
                        "nullable": false,
                        "example": "string"
                    },
                    "holdingIDShortHash": {
                        "type": "string",
                        "nullable": false,
                        "example": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": false,
                        "example": "2022-06-24T10:15:30Z"
                    }
                }
            },
...
            "StartFlowV1Response": {
                "type": "object",
                "required": ["flowStatus", "holdingIDShortHash", "timestamp"],
                "properties": {
                    "clientRequestId": {
                        "type": "string",
                        "nullable": true,
                        "example": "string"
                    },
                    "flowError": {
                        "$ref": "#/components/schemas/FlowV5Error"
                    },
                    "flowId": {
                        "type": "string",
                        "nullable": true,
                        "example": "string"
                    },
                    "flowResult": {
                        "type": "string",
                        "nullable": true,
                        "example": "string"
                    },
                    "flowStatus": {
                        "type": "string",
                        "nullable": false,
                        "example": "string"
                    },
                    "holdingIDShortHash": {
                        "type": "string",
                        "nullable": false,
                        "example": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": false,
                        "example": "2022-06-24T10:15:30Z"
                    }
                }
            },
...
            "FlowV5Error": {
                "type": "object",
                "required": ["message", "type"],
                "properties": {
                    "message": {
                        "type": "string",
                        "nullable": false,
                        "example": "string"
                    },
                    "type": {
                        "type": "string",
                        "nullable": false,
                        "example": "string"
                    }
                }
            }
        }
    },
    "paths": {
...
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/list-flows": {
            "post": {
                "operationId": "listFlowsV1",
                "x-hyperledger-cacti": {
                    "http": {
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/list-flows",
                        "verbLowerCase": "post"
                    }
                },
                "description": "Responds with a list of the flows on the Corda node.",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ListFlowsV1Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListFlowsV1Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/flow/{holdingIDShortHash}/{clientRequestID}": {
            "get": {
                "x-hyperledger-cactus": {
                    "http": {
                        "verbLowerCase": "get",
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/flow/{holdingIDShortHash}/{clientRequestID}"
                    }
                },
                "operationId": "flowStatusResponse",
                "summary": "This method gets the current status of the specified flow instance.",
                "parameters": [
                    {
                        "name": "holdingIDShortHash",
                        "in": "path",
                        "required": true,
                        "description": "Holding identity short hash",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "clientRequestID",
                        "in": "path",
                        "required": true,
                        "description": "Client request ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlowStatusV5Response"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/startFlow": {
            "post": {
                "x-hyperledger-cactus": {
                    "http": {
                        "verbLowerCase": "post",
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/startFlow"
                    }
                },
                "operationId": "startFlowV1",
                "summary": "This method starts a new instance for the specified flow for the specified holding identity.",
                "requestBody": {
                    "description": "Request body for starting a flow",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StartFlowV5Request"
                            }
                        }
                    }
                },
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StartFlowV1Response"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}